@@ -8,7 +8,7 @@
 #import "RCTView.h"
 #import "MPNativeAdRendering.h"
 #import "MPNativeAdConstants.h"
-#import <RCTTextView.h>
+#import "RCTTextView.h"
 #import <mopub-ios-sdk/MoPub.h>
 NS_ASSUME_NONNULL_BEGIN
 
